@@ -54,23 +56,26 @@
     "pug": "2.0.0-rc.2",
     "reflect-metadata": "0.1.10",
     "request-promise": "4.2.1",
+    "require-from-string": "^1.2.1",
     "rxjs": "5.4.1",
     "simple-encryptor": "^1.1.0",
+    "webpack": "3.4.1",
     "zone.js": "0.8.14"
   },
   "devDependencies": {
-    "@angular/animations": "4.2.5",
-    "@angular/cli": "1.2.3",
-    "@angular/common": "4.2.5",
-    "@angular/compiler": "4.2.5",
-    "@angular/compiler-cli": "4.2.5",
-    "@angular/core": "4.2.5",
-    "@angular/forms": "4.2.5",
-    "@angular/http": "4.2.5",
-    "@angular/language-service": "4.2.5",
-    "@angular/platform-browser": "4.2.5",
-    "@angular/platform-browser-dynamic": "4.2.5",
-    "@angular/router": "4.2.5",
+    "@angular/animations": "4.3.5",
+    "@angular/cli": "1.3.0-rc.3",
+    "@angular/common": "4.3.5",
+    "@angular/compiler": "4.3.5",
+    "@angular/compiler-cli": "4.3.5",
+    "@angular/core": "4.3.5",
+    "@angular/forms": "4.3.5",
+    "@angular/http": "4.3.5",
+    "@angular/language-service": "4.3.5",
+    "@angular/platform-browser": "4.3.5",
+    "@angular/platform-browser-dynamic": "4.3.5",
+    "@angular/platform-server": "4.3.5",
+    "@angular/router": "4.3.5",
     "@ngrx/db": "2.0.1",
     "@ngrx/effects": "^4.0.2",
     "@ngrx/router-store": "^4.0.2",
