@@ -1,6 +1,7 @@
-import { Injectable } from '@angular/core';
-import { CanActivate, CanActivateChild, Router, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
-import { UserService, UIService } from './index';
+import { Injectable } from '@angular/core'
+import { CanActivate, CanActivateChild, Router, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router'
+import { UserService } from './user.service'
+import { UIService } from './ui.service'
 
 @Injectable()
 export class AuthGuardService implements CanActivate, CanActivateChild {
@@ -12,24 +13,24 @@ export class AuthGuardService implements CanActivate, CanActivateChild {
 	) {}
 
   canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean {
-  	// console.log("can activate", this.userService.isLoggedIn());
+  	// console.log("can activate", this.userService.isLoggedIn())
     if (this.userService.isLoggedIn()) {
-    	return true;
+    	return true
     } else {
       // if (state.url && state.url !== '/' && state.url.indexOf("login") === -1 && state.url.indexOf("home") === -1) {
-      //   this.router.navigate(['/home/login'], { queryParams: { target: state.url }});
+      //   this.router.navigate(['/home/login'], { queryParams: { target: state.url }})
       // } else {
-      //   this.router.navigate(['/home']);
+      //   this.router.navigate(['/home'])
       // }
-      // this.ui.flash("Your login session has ended", "info");
+      // this.ui.flash("Your login session has ended", "info")
       
-      this.router.navigate(['/home/login'], { queryParams: { target: state.url }});
-    	return false;
+      this.router.navigate(['/home/login'], { queryParams: { target: state.url }})
+    	return false
     }
   }
 
   canActivateChild(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean {
-    return this.canActivate(route, state);
+    return this.canActivate(route, state)
   }
 
 }
