@@ -1,10 +1,11 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core'
+import { CommonModule } from '@angular/common'
+import { RouterModule } from '@angular/router'
 
-import { SharedModule } from '../shared/shared.module';
-import { SiteRoutingModule } from './site-routing.module';
+import { SharedModule } from '../shared/shared.module'
+import { SiteRoutingModule } from './site-routing.module'
 
-import { HomeComponent } from './components/home/home.component';
+import { HomeComponent } from './components/home/home.component'
 
 @NgModule({
   imports: [
